fix: add and set pagination in dummy data
